+++ src/axislike.rs
@@ -25,13 +25,14 @@ impl AxisDirection {
         }
     }
 
-    /// Checks if the given `value` represents an active input in this direction.
+    /// Checks if the given `value` represents an active input in this direction,
+    /// considering the specified `threshold`.
     #[must_use]
     #[inline]
-    pub fn is_active(&self, value: f32) -> bool {
+    pub fn is_active(&self, value: f32, threshold: f32) -> bool {
         match self {
-            Self::Negative => value < 0.0,
-            Self::Positive => value > 0.0,
+            Self::Negative => value < -threshold,
+            Self::Positive => value > threshold,
         }
     }
 }
@@ -151,11 +152,12 @@ impl DualAxisDirection {
         }
     }
 
-    /// Checks if the given `value` represents an active input in this direction.
+    /// Checks if the given `value` represents an active input in this direction,
+    /// considering the specified `threshold`.
     #[must_use]
     #[inline]
-    pub fn is_active(&self, value: Vec2) -> bool {
-        let component_along_axis = self.axis().get_value(value);
-        self.axis_direction().is_active(component_along_axis)
+    pub fn is_active(&self, value: Vec2, threshold: f32) -> bool {
+        let axis_value = self.axis().get_value(value);
+        self.axis_direction().is_active(axis_value, threshold)
     }
 }