+++ src/user_input/mouse.rs
@@ -1,12 +1,12 @@
 //! Mouse inputs
 
-use bevy::input::mouse::{MouseButtonInput, MouseMotion, MouseWheel};
+use bevy::input::mouse::{MouseButton, MouseButtonInput, MouseMotion, MouseWheel};
 use bevy::input::{ButtonInput, ButtonState};
-use bevy::prelude::{
-    Entity, Events, Gamepad, MouseButton, Reflect, Res, ResMut, Resource, Vec2, World,
-};
+use bevy::math::FloatOrd;
+use bevy::prelude::{Entity, Events, Gamepad, Reflect, Res, ResMut, Resource, Vec2, World};
 use leafwing_input_manager_macros::serde_typetag;
 use serde::{Deserialize, Serialize};
+use std::hash::{Hash, Hasher};
 
 use crate as leafwing_input_manager;
 use crate::axislike::{DualAxisDirection, DualAxisType};
@@ -112,22 +112,47 @@ impl Buttonlike for MouseButton {
 /// app.update();
 /// assert!(app.read_pressed(input));
 /// ```
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Reflect, Serialize, Deserialize)]
+#[derive(Debug, Clone, Copy, PartialEq, Reflect, Serialize, Deserialize)]
 #[must_use]
-pub struct MouseMoveDirection(pub DualAxisDirection);
+pub struct MouseMoveDirection {
+    /// The direction to monitor (up, down, left, or right).
+    pub direction: DualAxisDirection,
+
+    /// The threshold value for the direction to be considered pressed.
+    pub threshold: f32,
+}
 
 impl MouseMoveDirection {
+    /// Sets the `threshold` value.
+    #[inline]
+    pub const fn threshold(mut self, threshold: f32) -> Self {
+        self.threshold = threshold;
+        self
+    }
+
     /// Movement in the upward direction.
-    pub const UP: Self = Self(DualAxisDirection::Up);
+    pub const UP: Self = Self {
+        direction: DualAxisDirection::Up,
+        threshold: 0.0,
+    };
 
     /// Movement in the downward direction.
-    pub const DOWN: Self = Self(DualAxisDirection::Down);
+    pub const DOWN: Self = Self {
+        direction: DualAxisDirection::Down,
+        threshold: 0.0,
+    };
 
     /// Movement in the leftward direction.
-    pub const LEFT: Self = Self(DualAxisDirection::Left);
+    pub const LEFT: Self = Self {
+        direction: DualAxisDirection::Left,
+        threshold: 0.0,
+    };
 
     /// Movement in the rightward direction.
-    pub const RIGHT: Self = Self(DualAxisDirection::Right);
+    pub const RIGHT: Self = Self {
+        direction: DualAxisDirection::Right,
+        threshold: 0.0,
+    };
 }
 
 impl UserInput for MouseMoveDirection {
@@ -140,7 +165,7 @@ impl UserInput for MouseMoveDirection {
     /// [`MouseMoveDirection`] represents a simple virtual button.
     #[inline]
     fn decompose(&self) -> BasicInputs {
-        BasicInputs::Simple(Box::new(*self))
+        BasicInputs::Simple(Box::new((*self).threshold(0.0)))
     }
 }
 
@@ -151,7 +176,7 @@ impl Buttonlike for MouseMoveDirection {
     #[inline]
     fn pressed(&self, input_store: &CentralInputStore, _gamepad: Gamepad) -> bool {
         let mouse_movement = input_store.pair(&MouseMove::default());
-        self.0.is_active(mouse_movement)
+        self.direction.is_active(mouse_movement, self.threshold)
     }
 
     /// Sends a [`MouseMotion`] event with a magnitude of 1.0 in the direction defined by `self`.
@@ -159,7 +184,7 @@ impl Buttonlike for MouseMoveDirection {
         world
             .resource_mut::<Events<MouseMotion>>()
             .send(MouseMotion {
-                delta: self.0.full_active_value(),
+                delta: self.direction.full_active_value(),
             });
     }
 
@@ -168,6 +193,15 @@ impl Buttonlike for MouseMoveDirection {
     /// As mouse movement directions are determined based on the recent change in mouse position,
     /// no action other than waiting for the next frame is necessary to release the input.
     fn release(&self, _world: &mut World) {}
+}
+
+impl Eq for MouseMoveDirection {}
+
+impl Hash for MouseMoveDirection {
+    fn hash<H: Hasher>(&self, state: &mut H) {
+        self.direction.hash(state);
+        FloatOrd(self.threshold).hash(state);
+    }
 }
 
 /// Relative changes in position of mouse movement on a single axis (X or Y).
@@ -234,8 +268,14 @@ impl UserInput for MouseMoveAxis {
     #[inline]
     fn decompose(&self) -> BasicInputs {
         BasicInputs::Composite(vec![
-            Box::new(MouseMoveDirection(self.axis.negative())),
-            Box::new(MouseMoveDirection(self.axis.positive())),
+            Box::new(MouseMoveDirection {
+                direction: self.axis.negative(),
+                threshold: 0.0,
+            }),
+            Box::new(MouseMoveDirection {
+                direction: self.axis.positive(),
+                threshold: 0.0,
+            }),
         ])
     }
 }
@@ -422,22 +462,47 @@ impl WithDualAxisProcessingPipelineExt for MouseMove {
 /// app.update();
 /// assert!(app.read_pressed(input));
 /// ```
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Reflect, Serialize, Deserialize)]
+#[derive(Debug, Clone, Copy, PartialEq, Reflect, Serialize, Deserialize)]
 #[must_use]
-pub struct MouseScrollDirection(pub DualAxisDirection);
+pub struct MouseScrollDirection {
+    /// The direction to monitor (up, down, left, or right).
+    pub direction: DualAxisDirection,
+
+    /// The threshold value for the direction to be considered pressed.
+    pub threshold: f32,
+}
 
 impl MouseScrollDirection {
+    /// Sets the `threshold` value.
+    #[inline]
+    pub const fn threshold(mut self, threshold: f32) -> Self {
+        self.threshold = threshold;
+        self
+    }
+
     /// Movement in the upward direction.
-    pub const UP: Self = Self(DualAxisDirection::Up);
+    pub const UP: Self = Self {
+        direction: DualAxisDirection::Up,
+        threshold: 0.0,
+    };
 
     /// Movement in the downward direction.
-    pub const DOWN: Self = Self(DualAxisDirection::Down);
+    pub const DOWN: Self = Self {
+        direction: DualAxisDirection::Down,
+        threshold: 0.0,
+    };
 
     /// Movement in the leftward direction.
-    pub const LEFT: Self = Self(DualAxisDirection::Left);
+    pub const LEFT: Self = Self {
+        direction: DualAxisDirection::Left,
+        threshold: 0.0,
+    };
 
     /// Movement in the rightward direction.
-    pub const RIGHT: Self = Self(DualAxisDirection::Right);
+    pub const RIGHT: Self = Self {
+        direction: DualAxisDirection::Right,
+        threshold: 0.0,
+    };
 }
 
 impl UserInput for MouseScrollDirection {
@@ -450,7 +515,7 @@ impl UserInput for MouseScrollDirection {
     /// [`MouseScrollDirection`] represents a simple virtual button.
     #[inline]
     fn decompose(&self) -> BasicInputs {
-        BasicInputs::Simple(Box::new(*self))
+        BasicInputs::Simple(Box::new((*self).threshold(0.0)))
     }
 }
 
@@ -461,7 +526,7 @@ impl Buttonlike for MouseScrollDirection {
     #[inline]
     fn pressed(&self, input_store: &CentralInputStore, _gamepad: Gamepad) -> bool {
         let movement = input_store.pair(&MouseScroll::default());
-        self.0.is_active(movement)
+        self.direction.is_active(movement, self.threshold)
     }
 
     /// Sends a [`MouseWheel`] event with a magnitude of 1.0 px in the direction defined by `self`.
@@ -470,7 +535,7 @@ impl Buttonlike for MouseScrollDirection {
     ///
     /// The `window` field will be filled with a placeholder value.
     fn press(&self, world: &mut World) {
-        let vec = self.0.full_active_value();
+        let vec = self.direction.full_active_value();
 
         world.resource_mut::<Events<MouseWheel>>().send(MouseWheel {
             unit: bevy::input::mouse::MouseScrollUnit::Pixel,
@@ -485,6 +550,15 @@ impl Buttonlike for MouseScrollDirection {
     /// As mouse scroll directions are determined based on the recent change in mouse scrolling,
     /// no action other than waiting for the next frame is necessary to release the input.
     fn release(&self, _world: &mut World) {}
+}
+
+impl Eq for MouseScrollDirection {}
+
+impl Hash for MouseScrollDirection {
+    fn hash<H: Hasher>(&self, state: &mut H) {
+        self.direction.hash(state);
+        FloatOrd(self.threshold).hash(state);
+    }
 }
 
 /// Amount of mouse wheel scrolling on a single axis (X or Y).
@@ -551,8 +625,14 @@ impl UserInput for MouseScrollAxis {
     #[inline]
     fn decompose(&self) -> BasicInputs {
         BasicInputs::Composite(vec![
-            Box::new(MouseScrollDirection(self.axis.negative())),
-            Box::new(MouseScrollDirection(self.axis.positive())),
+            Box::new(MouseScrollDirection {
+                direction: self.axis.negative(),
+                threshold: 0.0,
+            }),
+            Box::new(MouseScrollDirection {
+                direction: self.axis.positive(),
+                threshold: 0.0,
+            }),
         ])
     }
 }